+++ src/pytask/__main__.py
@@ -1,0 +1,8 @@
+"""The pytask entry-point."""
+import sys
+
+import pytask
+
+
+if __name__ == "__main__":
+    sys.exit(pytask.cli())